refactor: move client to notion package
